refactor: 평점 별UI 로직을 UI모델로 분리
