+++ feature/library/src/main/java/com/into/websoso/feature/library/model/LibraryListItemModel.kt
@@ -1,6 +1,6 @@
 package com.into.websoso.feature.library.model
 
-import com.into.websoso.feature.library.util.formatDateRange
+import com.into.websoso.core.common.extensions.formatDateRange
 
 data class LibraryListItemModel(
     val novelId: Long,
@@ -16,4 +16,26 @@ data class LibraryListItemModel(
     val myFeeds: List<String>,
     val isInterest: Boolean,
     val formattedDateRange: String? = formatDateRange(startDate, endDate),
-)
+) {
+    val ratingStars: List<RatingStarType> = calculateRatingStars(userNovelRating)
+
+    private fun calculateRatingStars(rating: Float?): List<RatingStarType> {
+        if (rating == null) return emptyList()
+
+        val fullStar = rating.toInt()
+        val halfStar = (rating - fullStar) >= 0.5f
+        val emptyStar = 5 - fullStar - if (halfStar) 1 else 0
+
+        return buildList {
+            repeat(fullStar) { add(RatingStarType.FULL) }
+            if (halfStar) add(RatingStarType.HALF)
+            repeat(emptyStar) { add(RatingStarType.EMPTY) }
+        }
+    }
+}
+
+enum class RatingStarType {
+    FULL,
+    HALF,
+    EMPTY,
+}
